Feature/yahoo

+++ src/main/java/com/spxvol/www/model/Heatmap.java
@@ -1,6 +1,5 @@
 package com.spxvol.www.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Comparator;
 import java.util.List;
@@ -16,9 +15,9 @@ public class Heatmap {
 
 	private List<OptionQuote> chains;
 
-	private Map<LocalDate, Map<BigDecimal, List<OptionQuote>>> expirationMap;
+	private Map<LocalDate, Map<Double, List<OptionQuote>>> expirationMap;
 
-	private Map<BigDecimal, Map<LocalDate, List<OptionQuote>>> strikeMap;
+	private Map<Double, Map<LocalDate, List<OptionQuote>>> strikeMap;
 
 	private double minIV;
 
@@ -31,7 +30,7 @@ public class Heatmap {
 		chains.removeIf(o -> o.getDate().equals(underlying.getLastTradeTZ().toLocalDate()));
 
 		if (skipStrikes) {
-			Map<BigDecimal, List<OptionQuote>> datesByStrike = chains.stream()
+			Map<Double, List<OptionQuote>> datesByStrike = chains.stream()
 					.collect(Collectors.groupingBy(OptionQuote::getStrikePrice));
 			int maxDates = datesByStrike.values().stream().map(List::size).max(Comparator.naturalOrder()).get();
 			chains = datesByStrike.values().stream().filter(list -> list.size() == maxDates).flatMap(List::stream)
@@ -60,7 +59,7 @@ public class Heatmap {
 				.collect(Collectors.toList());
 	}
 
-	public List<BigDecimal> getStrikes() {
+	public List<Double> getStrikes() {
 		return chains.stream().map(OptionQuote::getStrikePrice).distinct().sorted().collect(Collectors.toList());
 	}
 	
@@ -77,7 +76,7 @@ public class Heatmap {
 		return String.format("hsl(%.2f, %.2f%%, %.2f%%)", h, s, l);
 	}
 	
-	public List<ColorfulValue> row(BigDecimal strike) {
+	public List<ColorfulValue> row(Double strike) {
 		Map<LocalDate, List<OptionQuote>> dates = strikeMap.get(strike);
 		
 		return expirationMap.keySet().stream().sorted().map(date -> {
